Moved _fake_click and updated tests accordingly
